refactor(app): share next_actions helpers
