+++ openspec/changes/refactor-app-next-actions/README.md
@@ -1,0 +1,3 @@
+# refactor-app-next-actions
+
+Refactor: share status next_actions ordering between CLI and MCP
