+++ src/app/mod.rs
@@ -1,1 +1,2 @@
+pub(crate) mod next_actions;
 pub(crate) mod preview_diff;
